Fixed convertion problem navi-195

+++ src/main/java/co/blastlab/serviceblbnavi/socket/info/in/FileListDetails.java
@@ -10,6 +10,6 @@ public class FileListDetails extends Info {
 	public String path;
 	public int fsize;
 	public byte[] md5;
-	public short fCRC;
+	public int fCRC;
 	public long createdUTC;
 }
